requirements.txt
flask
flask-admin
flask-sqlalchemy
flask-migrate
Flask-WTF
flask-mail
Flask-Login
email-validator
dynaconf
wtforms==2.3.3
Pillow